Untested support for FTDI D2XX driver
